Rewrote the "Wha to use it" sections
My view on those sections is to be a checklist for users to see if their situation or the way they use/write the input files matches what is expected by the templates.

For example for strings template, depending on if you're using dot-syntax in your Localizable.strings or not, you might want one template or the other, so that section is supposed to tell you which template to use depending on how you write your keys. In future other string templates we might add bullet points like "If you use underscores in your key names" or "if all your keys in your Localizable.strings are all full-uppercase", etc.